test(direct-control): extract restart lifecycle tests
Move restart and begin lifecycle package coverage out of the residual broad direct-control suite into packages/civ7-direct-control/test/restart-lifecycle.test.ts. Use a focused fake tuner for restart, begin, App UI snapshot, and Tuner health routing, then remove packages/civ7-direct-control/test/direct-control.test.ts.

Update the support modularization OpenSpec records to mark focused package-test atom coverage complete while keeping source extraction pending. This is local fake-tuner package proof only, not runtime proof.

Validation: bun run --cwd packages/civ7-direct-control test -- restart-lifecycle.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
